backend/requirements.txt
langchain
langchain-core
langchain-community
langchain-nvidia-ai-endpoints
langchain-ollama
chromadb
openai
websockets
python-dotenv
fastapi